+++ yaqd-fakes/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
 ## [Unreleased]
 
+## [2021.2.0]
+
 ### Added
 - new daemon "fake-spectrometer" implementing new has-mapping trait, see [YEP-311](https://yeps.yaq.fyi)
 
@@ -70,7 +72,8 @@ There are no actual code changes, this release is to update the release pipeline
 ### Added
 - initial release
 
-[Unreleased]: https://gitlab.com/yaq/yaq-python/-/compare/yaqd-fakes-2021.1.0...master
+[Unreleased]: https://gitlab.com/yaq/yaq-python/-/compare/yaqd-fakes-2021.2.0...master
+[2021.2.0]: https://gitlab.com/yaq/yaq-python/-/compare/yaqd-fakes-2021.1.0...yaqd-fakes-2021.2.0
 [2021.1.0]: https://gitlab.com/yaq/yaq-python/-/compare/yaqd-fakes-2020.10.0...yaqd-fakes-2021.1.0
 [2020.10.0]: https://gitlab.com/yaq/yaq-python/-/compare/yaqd-fakes-2020.09.1...yaqd-fakes-2020.10.0
 [2020.09.1]: https://gitlab.com/yaq/yaq-python/-/compare/yaqd-fakes-2020.07.4...yaqd-fakes-2020.09.1